refactor(event-pipeline): create scroll event processor
